original_requirements.txt
numpy==1.14.5
torchvision==0.2.2.post3
matplotlib
tensorflow==1.13.1
scipy
torch==1.1.0
vispy==0.5.3
opencv_python==4.1.0.25
opencv_contrib_python==4.1.0.25
Pillow==6.1.0
PyYAML==5.1.1
